Update generate_api_documentation.pl
Made slight change to the require line.  ----    require '../doxygen/perlmod/DoxyDocs.pm';   -----       This particular line was keeping the code from compiling on linux throwing a Doxygen error not being able to find the doxygen/perlmod/directory.

I also made a copy of the doxygen directory it was looking for,  into the /libraries directory.      Now the code compiles completely and works.